requirements.txt
Faker==0.8.4
PyGithub==1.35
elasticsearch-dsl==5.3.0
factory-boy==2.9.2
